Rename `int`s to `s32`/`u32`s and `float`s to `f32`s (#490)
* Misc. documentation

* whoops

* misc labelling

* Update interface_2.c

* Update interface_2.c

* i still dont know how doxygen works tbh

* still have no clue what im doing

* Label gMarioOnMerryGoRound

* more labelling

* even MORE misc labelling

* boo activation radius

* i forgot what i changed in this commit

* wow so this is what i changed, nice

* i forgot again

* audio defines.h

* .

* .

* Revert "."

This reverts commit eb4139abceefdceb82e3b3270eebb5799041dd8e.

idk what the fuck git did

* red i hate you

* label

* :morphon:

* Revert ":morphon:"

This reverts commit 22ac4ad5f2761ebebf24126581760513bd15d427.

:morphoff:

* Update boo.inc.c

* more labelling

* Update audio_defines.h

* Update yoshi.inc.c

* Update yoshi.inc.c

* rename respawner object more accurately

* label respawners more

* forgot to ctrl s

* moRE label

* build

* DOUBLE BAITED

* pray to god that this builds

* put exclamation box in exclamation box and label some

* please build

* make :ok:

* fix newline warnings

* why did this newline warning not show up before

* removed revo's name

* first 101 funcs changed

* next 101 funcs changed

* ok so far, next 101 funcs changed

* still ok, last 101 funcs changed, this has to be the fail

* beh -> bhv in behavior functions

* i forgot what i changed

* second wave

* ok i fixed 4 more

* never mind i'm a dumbass

* fix all the things

* Update README.md

* int -> s32

* fix

* float => f32

* did what simon said

* blegh

* blegh

* blegh

* blegh

* blegh

* Update profiler_utils.c

* blegh

+++ src/game/behaviors/hoot.inc.c
@@ -38,7 +38,7 @@ void HootFloorBounce(void)
     
     floorY = HootFindNextFloor(&sp1c, 0);
     if (floorY + 125.0f > o->oPosY) o->oPosY = floorY + 125.0f;
-    if ((f64)o->oMoveAnglePitch < -21845.3328) o->oMoveAnglePitch = -21845;
+    if (o->oMoveAnglePitch < -21845.3328) o->oMoveAnglePitch = -21845;
 }
 
 //sp30 = fastOscY
